pyproject.toml
[project]
name = "iron-counsel"
version = "0.1.0"
description = "Telegram RAG bot — Game of Thrones quotes via Gemma 3 + Firestore"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.8",
    "google-cloud-firestore==2.20.1",
    "httpx==0.28.1",
    "langchain==0.3.19",
    "langchain-core==0.3.40",
    "langchain-community>=0.3.0",
    "langchain-groq>=0.2.4",
    "pydantic-settings==2.8.1",
    "python-telegram-bot==21.10",
    "uvicorn[standard]==0.34.0",
    "tqdm>=4.67.3",
    "fastembed>=0.4.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
